Allow changing settings at the client level.
Added MongoClientOperations to handle client level operations and to allow
for changable settings.

JAVA-5259
