+++ config/spotbugs/exclude.xml
@@ -268,7 +268,7 @@
          see: https://github.com/Kotlin/kotlinx.coroutines/issues/3099
     -->
     <Match>
-        <Class name="com.mongodb.kotlin.client.coroutine.MongoClient"/>
+        <Class name="com.mongodb.kotlin.client.coroutine.MongoClientOperations"/>
         <Method name="startSession"/>
         <Bug pattern="NP_NULL_ON_SOME_PATH_FROM_RETURN_VALUE"/>
     </Match>
