+++ driver-reactive-streams/src/main/com/mongodb/reactivestreams/client/MongoClient.java
@@ -16,17 +16,12 @@
 
 package com.mongodb.reactivestreams.client;
 
-import com.mongodb.ClientSessionOptions;
 import com.mongodb.annotations.Immutable;
 import com.mongodb.connection.ClusterDescription;
 import com.mongodb.connection.ClusterSettings;
 import com.mongodb.event.ClusterListener;
-import org.bson.Document;
-import org.bson.conversions.Bson;
-import org.reactivestreams.Publisher;
 
 import java.io.Closeable;
-import java.util.List;
 
 /**
  * A client-side representation of a MongoDB cluster.  Instances can represent either a standalone MongoDB instance, a replica set,
@@ -39,193 +34,13 @@ import java.util.List;
  * @since 1.0
  */
 @Immutable
-public interface MongoClient extends Closeable {
-    /**
-     * Gets the database with the given name.
-     *
-     * @param name the name of the database
-     * @return the database
-     */
-    MongoDatabase getDatabase(String name);
+public interface MongoClient extends MongoClientOperations, Closeable {
 
     /**
      * Close the client, which will close all underlying cached resources, including, for example,
      * sockets and background monitoring threads.
      */
     void close();
-
-    /**
-     * Get a list of the database names
-     *
-     * @mongodb.driver.manual reference/commands/listDatabases List Databases
-     * @return an iterable containing all the names of all the databases
-     */
-    Publisher<String> listDatabaseNames();
-
-    /**
-     * Get a list of the database names
-     *
-     * @param clientSession the client session with which to associate this operation
-     * @mongodb.driver.manual reference/commands/listDatabases List Databases
-     * @return an iterable containing all the names of all the databases
-     *
-     * @mongodb.server.release 3.6
-     * @since 1.7
-     */
-    Publisher<String> listDatabaseNames(ClientSession clientSession);
-
-    /**
-     * Gets the list of databases
-     *
-     * @return the fluent list databases interface
-     */
-    ListDatabasesPublisher<Document> listDatabases();
-
-    /**
-     * Gets the list of databases
-     *
-     * @param clazz the class to cast the database documents to
-     * @param <TResult>   the type of the class to use instead of {@code Document}.
-     * @return the fluent list databases interface
-     */
-    <TResult> ListDatabasesPublisher<TResult> listDatabases(Class<TResult> clazz);
-
-    /**
-     * Gets the list of databases
-     *
-     * @param clientSession the client session with which to associate this operation
-     * @return the fluent list databases interface
-     * @mongodb.server.release 3.6
-     * @since 1.7
-     */
-    ListDatabasesPublisher<Document> listDatabases(ClientSession clientSession);
-
-    /**
-     * Gets the list of databases
-     *
-     * @param clientSession the client session with which to associate this operation
-     * @param clazz the class to cast the database documents to
-     * @param <TResult>   the type of the class to use instead of {@code Document}.
-     * @return the fluent list databases interface
-     * @mongodb.server.release 3.6
-     * @since 1.7
-     */
-    <TResult> ListDatabasesPublisher<TResult> listDatabases(ClientSession clientSession, Class<TResult> clazz);
-
-    /**
-     * Creates a change stream for this client.
-     *
-     * @return the change stream iterable
-     * @mongodb.driver.dochub core/changestreams Change Streams
-     * @since 1.9
-     * @mongodb.server.release 4.0
-     */
-    ChangeStreamPublisher<Document> watch();
-
-    /**
-     * Creates a change stream for this client.
-     *
-     * @param resultClass the class to decode each document into
-     * @param <TResult>   the target document type of the iterable.
-     * @return the change stream iterable
-     * @mongodb.driver.dochub core/changestreams Change Streams
-     * @since 1.9
-     * @mongodb.server.release 4.0
-     */
-    <TResult> ChangeStreamPublisher<TResult> watch(Class<TResult> resultClass);
-
-    /**
-     * Creates a change stream for this client.
-     *
-     * @param pipeline the aggregation pipeline to apply to the change stream.
-     * @return the change stream iterable
-     * @mongodb.driver.dochub core/changestreams Change Streams
-     * @since 1.9
-     * @mongodb.server.release 4.0
-     */
-    ChangeStreamPublisher<Document> watch(List<? extends Bson> pipeline);
-
-    /**
-     * Creates a change stream for this client.
-     *
-     * @param pipeline    the aggregation pipeline to apply to the change stream
-     * @param resultClass the class to decode each document into
-     * @param <TResult>   the target document type of the iterable.
-     * @return the change stream iterable
-     * @mongodb.driver.dochub core/changestreams Change Streams
-     * @since 1.9
-     * @mongodb.server.release 4.0
-     */
-    <TResult> ChangeStreamPublisher<TResult> watch(List<? extends Bson> pipeline, Class<TResult> resultClass);
-
-    /**
-     * Creates a change stream for this client.
-     *
-     * @param clientSession the client session with which to associate this operation
-     * @return the change stream iterable
-     * @since 1.9
-     * @mongodb.server.release 4.0
-     * @mongodb.driver.dochub core/changestreams Change Streams
-     */
-    ChangeStreamPublisher<Document> watch(ClientSession clientSession);
-
-    /**
-     * Creates a change stream for this client.
-     *
-     * @param clientSession the client session with which to associate this operation
-     * @param resultClass the class to decode each document into
-     * @param <TResult>   the target document type of the iterable.
-     * @return the change stream iterable
-     * @since 1.9
-     * @mongodb.server.release 4.0
-     * @mongodb.driver.dochub core/changestreams Change Streams
-     */
-    <TResult> ChangeStreamPublisher<TResult> watch(ClientSession clientSession, Class<TResult> resultClass);
-
-    /**
-     * Creates a change stream for this client.
-     *
-     * @param clientSession the client session with which to associate this operation
-     * @param pipeline the aggregation pipeline to apply to the change stream.
-     * @return the change stream iterable
-     * @since 1.9
-     * @mongodb.server.release 4.0
-     * @mongodb.driver.dochub core/changestreams Change Streams
-     */
-    ChangeStreamPublisher<Document> watch(ClientSession clientSession, List<? extends Bson> pipeline);
-
-    /**
-     * Creates a change stream for this client.
-     *
-     * @param clientSession the client session with which to associate this operation
-     * @param pipeline    the aggregation pipeline to apply to the change stream
-     * @param resultClass the class to decode each document into
-     * @param <TResult>   the target document type of the iterable.
-     * @return the change stream iterable
-     * @since 1.9
-     * @mongodb.server.release 4.0
-     * @mongodb.driver.dochub core/changestreams Change Streams
-     */
-    <TResult> ChangeStreamPublisher<TResult> watch(ClientSession clientSession, List<? extends Bson> pipeline, Class<TResult> resultClass);
-
-    /**
-     * Creates a client session.
-     *
-     * @return a publisher for the client session.
-     * @mongodb.server.release 3.6
-     * @since 1.9
-     */
-    Publisher<ClientSession> startSession();
-
-    /**
-     * Creates a client session.
-     *
-     * @param options the options for the client session
-     * @return a publisher for the client session.
-     * @mongodb.server.release 3.6
-     * @since 1.7
-     */
-    Publisher<ClientSession> startSession(ClientSessionOptions options);
 
     /**
      * Gets the current cluster description.
