+++ driver-scala/src/integration/scala/org/mongodb/scala/syncadapter/SyncMongoClientOperations.scala
@@ -1,0 +1,130 @@
+package org.mongodb.scala.syncadapter
+
+import com.mongodb.{ ClientSessionOptions, ReadConcern, ReadPreference, WriteConcern }
+import com.mongodb.client.{
+  ClientSession,
+  MongoClientOperations => JMongoClientOperations,
+  MongoDatabase => JMongoDatabase
+}
+import org.bson.Document
+import org.bson.codecs.configuration.CodecRegistry
+import org.bson.conversions.Bson
+import org.mongodb.scala.MongoClientOperations
+import org.mongodb.scala.bson.DefaultHelper.DefaultsTo
+
+import java.util.concurrent.TimeUnit
+import scala.collection.JavaConverters._
+import scala.concurrent.Await
+import scala.concurrent.duration.Duration
+import scala.reflect.ClassTag
+
+object SyncMongoClientOperations {
+
+  def apply(wrapped: MongoClientOperations): SyncMongoClientOperations = new SyncMongoClientOperations(wrapped)
+}
+
+class SyncMongoClientOperations(wrapped: MongoClientOperations) extends JMongoClientOperations {
+
+  override def getCodecRegistry: CodecRegistry = wrapped.codecRegistry
+
+  override def getReadPreference: ReadPreference = wrapped.readPreference
+
+  override def getWriteConcern: WriteConcern = wrapped.writeConcern
+
+  override def getReadConcern: ReadConcern = wrapped.readConcern
+
+  override def getTimeout(timeUnit: TimeUnit): java.lang.Long = {
+    val timeout = wrapped.timeout.map(d => timeUnit.convert(d.toMillis, TimeUnit.MILLISECONDS))
+    if (timeout.isDefined) timeout.get else null
+  }
+
+  override def withCodecRegistry(codecRegistry: CodecRegistry): JMongoClientOperations =
+    SyncMongoClientOperations(wrapped.withCodecRegistry(codecRegistry))
+
+  override def withReadPreference(readPreference: ReadPreference): JMongoClientOperations =
+    SyncMongoClientOperations(wrapped.withReadPreference(readPreference))
+
+  override def withWriteConcern(writeConcern: WriteConcern): JMongoClientOperations =
+    SyncMongoClientOperations(wrapped.withWriteConcern(writeConcern))
+
+  override def withReadConcern(readConcern: ReadConcern): JMongoClientOperations =
+    SyncMongoClientOperations(wrapped.withReadConcern(readConcern))
+
+  override def withTimeout(timeout: Long, timeUnit: TimeUnit): JMongoClientOperations =
+    SyncMongoClientOperations(wrapped.withTimeout(Duration(timeout, timeUnit)))
+
+  override def getDatabase(databaseName: String): JMongoDatabase =
+    SyncMongoDatabase(wrapped.getDatabase(databaseName))
+
+  override def startSession: ClientSession =
+    SyncClientSession(Await.result(wrapped.startSession().head(), WAIT_DURATION), this)
+
+  override def startSession(options: ClientSessionOptions): ClientSession =
+    SyncClientSession(Await.result(wrapped.startSession(options).head(), WAIT_DURATION), this)
+
+  override def listDatabaseNames = throw new UnsupportedOperationException
+
+  override def listDatabaseNames(clientSession: ClientSession) = throw new UnsupportedOperationException
+
+  override def listDatabases = new SyncListDatabasesIterable[Document](wrapped.listDatabases[Document]())
+
+  override def listDatabases(clientSession: ClientSession) = throw new UnsupportedOperationException
+
+  override def listDatabases[TResult](resultClass: Class[TResult]) =
+    new SyncListDatabasesIterable[TResult](
+      wrapped.listDatabases[TResult]()(
+        DefaultsTo.overrideDefault[TResult, org.mongodb.scala.Document],
+        ClassTag(resultClass)
+      )
+    )
+
+  override def listDatabases[TResult](clientSession: ClientSession, resultClass: Class[TResult]) =
+    throw new UnsupportedOperationException
+
+  override def watch = new SyncChangeStreamIterable[Document](wrapped.watch[Document]())
+
+  override def watch[TResult](resultClass: Class[TResult]) =
+    new SyncChangeStreamIterable[TResult](
+      wrapped.watch[TResult]()(DefaultsTo.overrideDefault[TResult, org.mongodb.scala.Document], ClassTag(resultClass))
+    )
+
+  override def watch(pipeline: java.util.List[_ <: Bson]) =
+    new SyncChangeStreamIterable[Document](wrapped.watch[Document](pipeline.asScala.toSeq))
+
+  override def watch[TResult](pipeline: java.util.List[_ <: Bson], resultClass: Class[TResult]) =
+    new SyncChangeStreamIterable[TResult](
+      wrapped.watch[TResult](pipeline.asScala.toSeq)(
+        DefaultsTo.overrideDefault[TResult, org.mongodb.scala.Document],
+        ClassTag(resultClass)
+      )
+    )
+
+  override def watch(clientSession: ClientSession) =
+    new SyncChangeStreamIterable[Document](wrapped.watch[Document](unwrap(clientSession)))
+
+  override def watch[TResult](clientSession: ClientSession, resultClass: Class[TResult]) =
+    new SyncChangeStreamIterable[TResult](
+      wrapped.watch(unwrap(clientSession))(
+        DefaultsTo.overrideDefault[TResult, org.mongodb.scala.Document],
+        ClassTag(resultClass)
+      )
+    )
+
+  override def watch(clientSession: ClientSession, pipeline: java.util.List[_ <: Bson]) =
+    new SyncChangeStreamIterable[Document](wrapped.watch[Document](unwrap(clientSession), pipeline.asScala.toSeq))
+
+  override def watch[TResult](
+      clientSession: ClientSession,
+      pipeline: java.util.List[_ <: Bson],
+      resultClass: Class[TResult]
+  ) =
+    new SyncChangeStreamIterable[TResult](
+      wrapped.watch[TResult](unwrap(clientSession), pipeline.asScala.toSeq)(
+        DefaultsTo.overrideDefault[TResult, org.mongodb.scala.Document],
+        ClassTag(resultClass)
+      )
+    )
+
+  private def unwrap(clientSession: ClientSession): org.mongodb.scala.ClientSession =
+    clientSession.asInstanceOf[SyncClientSession].wrapped
+}
