+++ driver-scala/src/test/scala/org/mongodb/scala/ApiAliasAndCompanionSpec.scala
@@ -87,6 +87,7 @@ class ApiAliasAndCompanionSpec extends BaseSpec {
       "AggregatePrimer",
       "RemovePrimer",
       "SyncMongoClient",
+      "SyncMongoClientOperations",
       "SyncGridFSBucket",
       "SyncMongoDatabase",
       "SyncClientEncryption"
